moved pswd reset&forgot to password.js, chnged ejs
