+++ views/forgotPassword.ejs
@@ -23,7 +23,7 @@
                                     <%= messages.error %>
                                 </div>
                                 <% } %>
-                                    <form action="/forgot-password" method="POST">
+                                    <form action="/password/forgot" method="POST">
                                         <div class="form-group">
                                             <label for="email">Email</label>
                                             <input type="email" name="email" class="form-control" id="email" aria-describedby="emailHelp">
