+++ views/partial/_sitemap.ejs
@@ -1,9 +1,9 @@
 <div class="site-links">
     <h3>bunch of links to help navigate through the site</h3>
     <a href="../create">create</a>
-    <a href="../forgot-password">forgot password</a>
+    <a href="../password/forgot">forgot password</a>
     <a href="../">index</a>
     <a href="../account/login">login</a>
     <a href="../registration">registration</a>
-    <a href="../reset-password">resetPassword</a>
+    <a href="../password/reset">resetPassword</a>
 </div>
